Conditional compilation with macros
Version-specific source directories for different Spark versions worked
well so far, but sometimes it is too much of a hassle. Macros can be
used to conditionally compile a part of code depending on the Spark
version, which is useful when it is not easy or requires duplicating a
lot of code to separate the part into multiple source directories.

For Spark 3+, Jackson annotations are not required for StructType. Using
the macros, we can remove unnecessary steps for Spark 3+.